+++ app/helpers/subjects_helper.rb
@@ -1,11 +1,4 @@
 # frozen_string_literal: true
 
 module SubjectsHelper
-  def status_helper(subject, long = false)
-    content_tag(  :span, (long ? subject.status : subject.status[0]),
-                  class: "label label-#{subject.status == 'valid' ? 'success' : 'info'}",
-                  rel: 'tooltip',
-                  data: { placement: 'left' },
-                  title: "#{subject.status unless long}" )
-  end
 end