+++ app/models/subject.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
 class Subject < ActiveRecord::Base
-  STATUS = %w(valid test).collect { |i| [i, i] }
-
   # Concerns
   include Searchable, Deletable
 
@@ -50,7 +48,7 @@ class Subject < ActiveRecord::Base
     current_user.all_subjects.where(id: id).count == 1
   end
 
-  def self.first_or_create_with_defaults(project, subject_code, acrostic, user, default_site, default_status)
+  def self.first_or_create_with_defaults(project, subject_code, acrostic, user, default_site)
     # (1) Find existing subject...
     subject = project.subjects.where(subject_code: subject_code).first
     return subject if subject
@@ -58,10 +56,9 @@ class Subject < ActiveRecord::Base
     site = project.sites.find_by_id(project.site_id_with_prefix(subject_code))
     if site
       default_site = site
-      default_status = 'valid'
     end
 
-    subject = project.subjects.where(subject_code: subject_code).first_or_create(acrostic: acrostic, user_id: user.id, site_id: default_site.id, status: default_status)
+    subject = project.subjects.where(subject_code: subject_code).first_or_create(acrostic: acrostic, user_id: user.id, site_id: default_site.id)
     subject
   end